use ember-infinity to load search cards on demand
