+++ app/controllers/cards.js
@@ -10,9 +10,6 @@ export default Ember.ArrayController.extend({
   /** @property {Array} array container for legality query params*/
   legalities: [],
 
-  /** @property {Number} the page number of cards to retrieve*/
-  page: 0,
-
   /** @property {String} the current search term */
   nameSearch: ''
 });
