+++ app/styles/components/infinity-loader.scss
@@ -1,0 +1,9 @@
+.infinity-loader {
+  text-align: center;
+  margin: 15px;
+
+  i {
+    font-size: 24px;
+    line-height: 24px;
+  }
+}
