+++ app/templates/application.hbs
@@ -66,10 +66,10 @@
         </h2>
       </div>
     {{/paper-toolbar}}
-    {{#paper-content}}
-    <div class="container-fluid">
-      {{outlet}}
-    </div>
+    {{#paper-content id="infinity-window"}}
+      <div class="container-fluid">
+        {{outlet}}
+      </div>
     {{/paper-content}}
   {{/paper-content}}
