+++ app/templates/components/infinity-loader.hbs
@@ -1,0 +1,5 @@
+{{#if infinityModel}}
+  {{#unless infinityModel.reachedInfinity}}
+    {{fa-icon "refresh" spin=true}}
+  {{/unless}}
+{{/if}}
