+++ package.json
@@ -38,9 +38,10 @@
     "ember-cli-qunit": "0.3.9",
     "ember-cli-selectize": "0.3.0",
     "ember-cli-uglify": "1.0.1",
+    "ember-cli-version-checker": "^1.0.2",
     "ember-data": "1.0.0-beta.15",
     "ember-export-application-global": "^1.0.2",
-    "ember-infinite-scroll": "^0.1.4",
+    "ember-infinity": "0.0.5",
     "ember-paper": "0.0.21",
     "emberfire": "1.4.3",
     "express": "^4.8.5",
